+++ ml-algorithms/src/main/java/org/opensearch/ml/engine/algorithms/sparse_encoding/SparseEncodingModel.java
@@ -35,22 +35,6 @@ import java.util.Map;
 public class SparseEncodingModel extends DLModel {
 
     @Override
-    public ModelTensorOutput predict(String modelId, MLInput mlInput) throws TranslateException {
-        MLInputDataset inputDataSet = mlInput.getInputDataset();
-        List<ModelTensors> tensorOutputs = new ArrayList<>();
-        Output output;
-        TextDocsInputDataSet textDocsInput = (TextDocsInputDataSet) inputDataSet;
-        ModelResultFilter resultFilter = textDocsInput.getResultFilter();
-        for (String doc : textDocsInput.getDocs()) {
-            Input input = new Input();
-            input.add(doc);
-            output = getPredictor().predict(input);
-            tensorOutputs.add(parseModelTensorOutput(output, resultFilter));
-        }
-        return new ModelTensorOutput(tensorOutputs);
-    }
-
-    @Override
     public Translator<Input, Output> getTranslator(String engine, MLModelConfig modelConfig) {
         return new SparseEncodingTranslator();
     }
@@ -58,36 +42,6 @@ public class SparseEncodingModel extends DLModel {
     @Override
     public TranslatorFactory getTranslatorFactory(String engine, MLModelConfig modelConfig) {
         return null;
-    }
-
-    @Override
-    public Map<String, Object> getArguments(MLModelConfig modelConfig) {
-        Map<String, Object> arguments = new HashMap<>();
-        if (modelConfig == null){
-            return arguments;
-        }
-        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
-        Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
-        if (modelMaxLength != null) {
-            arguments.put("modelMaxLength", modelMaxLength);
-        }
-        return arguments;
-    }
-
-    @Override
-    public void warmUp(Predictor predictor, String modelId, MLModelConfig modelConfig) throws TranslateException {
-        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
-        String warmUpSentence = "warm up sentence";
-        if (modelConfig  != null) {
-            Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
-            if (modelMaxLength != null) {
-                warmUpSentence = "sentence ".repeat(modelMaxLength);
-            }
-        }
-        // First request takes longer time. Predict once to warm up model.
-        Input input = new Input();
-        input.add(warmUpSentence);
-        predictor.predict(input);
     }
 
 }
